# fix: reuse comparators on subset