@wraithgar 2023-04-07
I don't *think* this overlaps with #538 in any way but am making a note to myself to triple check.