@wraithgar 2023-04-10
This is a tiny startup tradeoff for a larger runtime improvement.